style: remove unnecessary int32 conversions in ping-pong tests

+++ pkg/plugin/plugin_test.go
@@ -1949,7 +1949,7 @@ func TestSetHTTPRouteWeightPingPong(t *testing.T) {
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PingServiceName), updated.Spec.Rules[0].BackendRefs[0].Name)
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PongServiceName), updated.Spec.Rules[0].BackendRefs[1].Name)
 		// ping is stable → restWeight; pong is canary → desiredWeight
-		assert.Equal(t, int32(100-desiredWeight), *updated.Spec.Rules[0].BackendRefs[0].Weight)
+		assert.Equal(t, 100-desiredWeight, *updated.Spec.Rules[0].BackendRefs[0].Weight)
 		assert.Equal(t, desiredWeight, *updated.Spec.Rules[0].BackendRefs[1].Weight)
 	})
 
@@ -1974,7 +1974,7 @@ func TestSetHTTPRouteWeightPingPong(t *testing.T) {
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PongServiceName), updated.Spec.Rules[0].BackendRefs[1].Name)
 		// pong is stable → restWeight; ping is canary → desiredWeight
 		assert.Equal(t, desiredWeight, *updated.Spec.Rules[0].BackendRefs[0].Weight)
-		assert.Equal(t, int32(100-desiredWeight), *updated.Spec.Rules[0].BackendRefs[1].Weight)
+		assert.Equal(t, 100-desiredWeight, *updated.Spec.Rules[0].BackendRefs[1].Weight)
 	})
 }
 
@@ -2001,7 +2001,7 @@ func TestSetGRPCRouteWeightPingPong(t *testing.T) {
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PingServiceName), updated.Spec.Rules[0].BackendRefs[0].Name)
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PongServiceName), updated.Spec.Rules[0].BackendRefs[1].Name)
 		// ping is stable → restWeight; pong is canary → desiredWeight
-		assert.Equal(t, int32(100-desiredWeight), *updated.Spec.Rules[0].BackendRefs[0].Weight)
+		assert.Equal(t, 100-desiredWeight, *updated.Spec.Rules[0].BackendRefs[0].Weight)
 		assert.Equal(t, desiredWeight, *updated.Spec.Rules[0].BackendRefs[1].Weight)
 	})
 
@@ -2026,7 +2026,7 @@ func TestSetGRPCRouteWeightPingPong(t *testing.T) {
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PongServiceName), updated.Spec.Rules[0].BackendRefs[1].Name)
 		// pong is stable → restWeight; ping is canary → desiredWeight
 		assert.Equal(t, desiredWeight, *updated.Spec.Rules[0].BackendRefs[0].Weight)
-		assert.Equal(t, int32(100-desiredWeight), *updated.Spec.Rules[0].BackendRefs[1].Weight)
+		assert.Equal(t, 100-desiredWeight, *updated.Spec.Rules[0].BackendRefs[1].Weight)
 	})
 }
 
@@ -2053,7 +2053,7 @@ func TestSetTCPRouteWeightPingPong(t *testing.T) {
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PingServiceName), updated.Spec.Rules[0].BackendRefs[0].Name)
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PongServiceName), updated.Spec.Rules[0].BackendRefs[1].Name)
 		// ping is stable → restWeight; pong is canary → desiredWeight
-		assert.Equal(t, int32(100-desiredWeight), *updated.Spec.Rules[0].BackendRefs[0].Weight)
+		assert.Equal(t, 100-desiredWeight, *updated.Spec.Rules[0].BackendRefs[0].Weight)
 		assert.Equal(t, desiredWeight, *updated.Spec.Rules[0].BackendRefs[1].Weight)
 	})
 
@@ -2078,7 +2078,7 @@ func TestSetTCPRouteWeightPingPong(t *testing.T) {
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PongServiceName), updated.Spec.Rules[0].BackendRefs[1].Name)
 		// pong is stable → restWeight; ping is canary → desiredWeight
 		assert.Equal(t, desiredWeight, *updated.Spec.Rules[0].BackendRefs[0].Weight)
-		assert.Equal(t, int32(100-desiredWeight), *updated.Spec.Rules[0].BackendRefs[1].Weight)
+		assert.Equal(t, 100-desiredWeight, *updated.Spec.Rules[0].BackendRefs[1].Weight)
 	})
 }
 
@@ -2105,7 +2105,7 @@ func TestSetTLSRouteWeightPingPong(t *testing.T) {
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PingServiceName), updated.Spec.Rules[0].BackendRefs[0].Name)
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PongServiceName), updated.Spec.Rules[0].BackendRefs[1].Name)
 		// ping is stable → restWeight; pong is canary → desiredWeight
-		assert.Equal(t, int32(100-desiredWeight), *updated.Spec.Rules[0].BackendRefs[0].Weight)
+		assert.Equal(t, 100-desiredWeight, *updated.Spec.Rules[0].BackendRefs[0].Weight)
 		assert.Equal(t, desiredWeight, *updated.Spec.Rules[0].BackendRefs[1].Weight)
 	})
 
@@ -2130,6 +2130,6 @@ func TestSetTLSRouteWeightPingPong(t *testing.T) {
 		assert.Equal(t, gatewayv1.ObjectName(mocks.PongServiceName), updated.Spec.Rules[0].BackendRefs[1].Name)
 		// pong is stable → restWeight; ping is canary → desiredWeight
 		assert.Equal(t, desiredWeight, *updated.Spec.Rules[0].BackendRefs[0].Weight)
-		assert.Equal(t, int32(100-desiredWeight), *updated.Spec.Rules[0].BackendRefs[1].Weight)
+		assert.Equal(t, 100-desiredWeight, *updated.Spec.Rules[0].BackendRefs[1].Weight)
 	})
 }
